Update imports based on rename
